requirements.txt
# Core dependencies
fastapi>=0.104.1
python-multipart>=0.0.6
uvicorn>=0.24.0

# File processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"

# Text processing
nltk>=3.8.1
spacy>=3.7.2
pymorphy2>=0.9.1
beautifulsoup4>=4.12.2

# Data processing
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
regex>=2023.10.3

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1